requirements.txt
starlette~=0.31.1
uvicorn==0.23.2
Jinja2~=3.1.2
pymongo~=4.5.0
motor==3.3.1
requests==2.31.0
cryptography==41.0.4
wheel==0.41.2
google-auth==2.23.2
google~=3.0.0
lxml~=4.9.3
python-dotenv==1.0.0
argon2-cffi~=23.1.0
